Move function from test to utils module
