+++ nucliadb/nucliadb/writer/tests/test_fields.py
@@ -18,8 +18,6 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 
-import hashlib
-from base64 import b64encode
 from copy import deepcopy
 from datetime import datetime
 from os.path import dirname
@@ -34,21 +32,8 @@ from nucliadb.writer.api.v1.router import (
     RESOURCES_PREFIX,
     RSLUG_PREFIX,
 )
+from nucliadb.writer.tests.utils import load_file_as_FileB64_payload
 from nucliadb_models.resource import NucliaDBRoles
-
-
-def load_file_as_FileB64_payload(f: str, content_type: str) -> dict:
-    file_location = f"{dirname(__file__)}/{f}"
-    filename = f.split("/")[-1]
-    data = b64encode(open(file_location, "rb").read())
-
-    return {
-        "filename": filename,
-        "content_type": content_type,
-        "payload": data.decode("utf-8"),
-        "md5": hashlib.md5(data).hexdigest(),
-    }
-
 
 TEST_FILE = {f"{dirname(__file__)}/orm/"}
 TEST_TEXT_PAYLOAD = {"body": "test1", "format": "PLAIN"}
